refactor(latex): reorganize LaTeX-related code into separate modules
- Move LaTeX-related functions from figUtils and texUtils to dedicated modules
- Create new modules: extractFigure, latexdiff, latexindent, texcount, tikzPicture
- Standardize error handling and logging across LaTeX modules
- Move latex_to_pdf function from tikzpicture.py to tex_tools.py in Python package
- Aligned channels and logging across all LaTeX-related modules
