+++ coauthor-for-vs-code/src/latex/latexdiff.ts
@@ -2,13 +2,7 @@ import * as vscode from 'vscode';
 import * as path from 'path';
 import { promisify } from 'util';
 import * as cp from 'child_process';
-import {
-  getWorkspacePath,
-  readFile,
-  writeFile,
-  deleteFile,
-  fileExists,
-} from './fileUtils';
+import { getWorkspacePath, readFile, writeFile } from '../utils/fileUtils';
 import {
   debug,
   info,
@@ -16,11 +10,10 @@ import {
   error,
   initializeLogging,
 } from '../logger/logUtils';
-import { sync as globSync } from 'glob';
 
 const execAsync = promisify(cp.exec);
 
-const CHANNEL = 'TexUtils';
+const CHANNEL = 'LaTeX';
 initializeLogging(CHANNEL);
 
 async function processDiffFile(diffFileName: string): Promise<void> {
@@ -296,200 +289,4 @@ export async function runLatexDiffVCMultiple(
   }
 
   info(CHANNEL, 'All LaTeX diff operations completed');
-}
-
-export async function runLatexIndent(filePath: string): Promise<boolean> {
-  try {
-    const workspacePath = getWorkspacePath();
-    if (!workspacePath) {
-      throw new Error('No workspace path found');
-    }
-
-    // Get latexindent config from settings
-    const config = vscode.workspace.getConfiguration('coauthor.latex');
-    const latexindentConfig = config.get<string>('latexindentConfig');
-
-    // Build command array - note we're using -w (overwrite) and -s (silent)
-    const command = ['latexindent', '-w', '-s'];
-    if (latexindentConfig) {
-      command.push(`-l=${latexindentConfig}`);
-    }
-    command.push(`"${filePath}"`);
-
-    debug(CHANNEL, `Running command: ${command.join(' ')}`);
-
-    // Execute latexindent from workspace root
-    const { stdout, stderr } = await execAsync(command.join(' '), {
-      cwd: workspacePath,
-    });
-
-    if (stderr && stderr.trim()) {
-      warn(CHANNEL, `Latexindent stderr: ${stderr}`);
-    }
-
-    // Wait a moment for the file system to stabilize
-    await new Promise((resolve) => setTimeout(resolve, 100));
-
-    // Setup cleanup patterns relative to workspace
-    const fileBaseName = path.basename(filePath, '.tex');
-    const fileDir = path.dirname(filePath);
-
-    // Get all backup files matching the patterns, relative to workspace
-    const backupPatterns = [
-      path.join(fileDir, `${fileBaseName}.tex.bak*`),
-      path.join(fileDir, `${fileBaseName}.tex.bak`),
-      path.join(fileDir, `${fileBaseName}.bak*`),
-      path.join(fileDir, `${fileBaseName}.bak`),
-    ];
-
-    // Clean up backup files from workspace directory
-    for (const pattern of backupPatterns) {
-      const backupFiles = globSync(pattern, {
-        cwd: workspacePath,
-        absolute: false,
-      });
-
-      for (const backupFile of backupFiles) {
-        try {
-          await deleteFile(backupFile);
-          debug(CHANNEL, `Removed backup file: ${backupFile}`);
-        } catch (err) {
-          warn(CHANNEL, `Error removing backup file ${backupFile}: ${err}`);
-        }
-      }
-    }
-
-    // Clean up indent.log
-    const indentLogPath = path.join(path.dirname(filePath), 'indent.log');
-    try {
-      await deleteFile(indentLogPath);
-      debug(CHANNEL, 'Removed indent.log');
-    } catch (err) {
-      // Ignore error if indent.log doesn't exist
-      warn(CHANNEL, `Error removing indent.log: ${err}`);
-    }
-
-    info(CHANNEL, `Indented ${filePath}`);
-    return true;
-  } catch (err) {
-    error(
-      CHANNEL,
-      `Error running LaTeX indent: ${err instanceof Error ? err.message : String(err)}`,
-    );
-    return false;
-  }
-}
-
-/**
- * Compile a LaTeX file to PDF
- * @param texFile Path to the LaTeX file
- * @returns Promise<boolean> True if compilation succeeded
- */
-export async function compileLatexToPdf(texFile: string): Promise<boolean> {
-  try {
-    const workspacePath = getWorkspacePath();
-    if (!workspacePath) {
-      throw new Error('No workspace path found');
-    }
-
-    const outputDirectory = path.dirname(texFile);
-    const command = [
-      'pdflatex',
-      '-interaction=nonstopmode',
-      `-output-directory="${outputDirectory}"`,
-      `"${texFile}"`,
-    ].join(' ');
-
-    debug(CHANNEL, `Running command: ${command}`);
-    const { stdout, stderr } = await execAsync(command, {
-      cwd: workspacePath,
-    });
-
-    if (stderr && stderr.trim()) {
-      warn(CHANNEL, `pdflatex stderr: ${stderr}`);
-    }
-
-    info(CHANNEL, `Successfully compiled ${texFile}`);
-    return true;
-  } catch (err) {
-    error(
-      CHANNEL,
-      `Error compiling LaTeX: ${err instanceof Error ? err.message : String(err)}`,
-    );
-    return false;
-  }
-}
-
-/**
- * Get full statistics for LaTeX documents using the texcount Perl script
- * @param filePaths Single file path or array of file paths
- * @param merge Whether to merge included files in the count
- * @returns Promise<string | null> String containing full texcount output for all files, or null if an error occurred
- */
-export async function getTexCount(
-  filePaths: string | string[],
-  merge: boolean = false,
-): Promise<string | null> {
-  try {
-    const workspacePath = getWorkspacePath();
-    if (!workspacePath) {
-      throw new Error('No workspace path found');
-    }
-
-    // Convert single path to array
-    const paths = Array.isArray(filePaths) ? filePaths : [filePaths];
-    const allOutputs: string[] = [];
-
-    for (const filePath of paths) {
-      if (!(await fileExists(filePath))) {
-        warn(CHANNEL, `Warning: File ${filePath} does not exist.`);
-        continue;
-      }
-
-      if (!filePath.endsWith('.tex')) {
-        warn(CHANNEL, `Error: File ${filePath} is not a LaTeX file. Skipping.`);
-        continue;
-      }
-
-      const command = ['texcount'];
-      if (merge) {
-        command.push('-merge');
-      }
-      command.push(`"${filePath}"`);
-
-      debug(CHANNEL, `Running command: ${command.join(' ')}`);
-      try {
-        const { stdout, stderr } = await execAsync(command.join(' '), {
-          cwd: workspacePath,
-        });
-
-        if (stderr && stderr.trim()) {
-          warn(CHANNEL, `texcount stderr: ${stderr}`);
-        }
-
-        allOutputs.push(`Tex Count Results for ${filePath}:\n${stdout}`);
-        debug(CHANNEL, `Successfully counted ${filePath}`);
-      } catch (err) {
-        error(CHANNEL, `Error getting tex count for ${filePath}`);
-        error(
-          CHANNEL,
-          `Error: ${err instanceof Error ? err.message : String(err)}`,
-        );
-      }
-    }
-
-    if (allOutputs.length > 0) {
-      const combinedOutput = allOutputs.join('\n\n');
-      info(CHANNEL, `Combined Tex Count Results:\n${combinedOutput}`);
-      return combinedOutput;
-    }
-
-    return null;
-  } catch (err) {
-    error(
-      CHANNEL,
-      `Error in getTexCount: ${err instanceof Error ? err.message : String(err)}`,
-    );
-    return null;
-  }
 }
