+++ coauthor/latex/tex_tools.py
@@ -1,3 +1,4 @@
+import os
 import subprocess
 
 from ..utils.file import write_file
@@ -35,3 +36,21 @@ def run_external_command(
             error_message += f"\nStderr:\n{truncate_output(e.stderr)}"
         logger.error("\n" + error_message)
         return False, None, error_message
+
+
+def compile_latex_to_pdf(tex_file: str) -> bool:
+    """Compile a LaTeX file to PDF using pdflatex."""
+    output_directory = os.path.dirname(tex_file)
+    command = ["pdflatex", "-interaction=nonstopmode", "-output-directory=" + output_directory, tex_file]
+
+    try:
+        success, stdout, stderr = run_external_command(command, capture_output=True)
+        if success:
+            logger.info(f"Compiled {tex_file} successfully.")
+            return True
+        else:
+            logger.error(f"Error compiling {tex_file}")
+            return False
+    except ValueError as e:
+        logger.error(f"Error compiling {tex_file}: {str(e)}")
+        return False
